done with api endpoints for adding and removing tms

+++ src/trashmonsters/models.py
@@ -5,12 +5,12 @@ class Counters(models.Model):
     Current_TM_ID = models.SmallIntegerField()
     Current_Team_ID = models.SmallIntegerField()
 
-    # Makes it so there can only be one record for Counterspy
+    # Makes it so there can only be one record for Counters
     def has_add_permission(self, *args, **kwargs):
         return not Counters.objects.exists()
 
 class TrashMonsters(models.Model):
-    TM_ID = models.SmallIntegerField(primary_key=True)
+    TM_ID = models.AutoField(primary_key=True)
     Latitude = models.FloatField()
     Longitude = models.FloatField()
 
@@ -18,6 +18,3 @@ class TrashMonsters(models.Model):
     # ToString method for debug
     def __str__(self):
         return ("TM's ID: %d, Lat: %f, Long: %d,"%(self.TM_ID, self.Latitude, self.Longitude))
-
-
-
